+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -74,7 +74,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-05-20T13:20:42.306Z")
+    created: Optional[datetime] = parse_datetime("2025-05-20T13:37:36.239Z")
     r"""The date and time the resource was created"""