+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -78,5 +78,5 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-05-20T13:20:39.505Z")
+    updated: Optional[datetime] = parse_datetime("2025-05-20T13:37:33.567Z")
     r"""The date and time the resource was last updated"""